requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
asttokens @ file:///croot/asttokens_1743630435401/work
attrs==25.4.0
av==16.1.0
bitsandbytes==0.49.2
certifi==2026.2.25
charset-normalizer==3.4.4
comm @ file:///home/task_176311912503377/conda-bld/comm_1763119154572/work
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
cut-cross-entropy==25.1.1
datasets==4.3.0
debugpy @ file:///home/task_176242086745011/conda-bld/debugpy_1762421636577/work
decorator @ file:///home/task_175734099691499/conda-bld/decorator_1757341236139/work
diffusers==0.36.0
dill==0.4.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
executing @ file:///home/task_175706097783255/conda-bld/executing_1757061230504/work
filelock==3.25.0
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.3.2
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
ipykernel @ file:///home/task_176294602487836/conda-bld/ipykernel_1762946074244/work
ipython @ file:///home/task_176278907235969/conda-bld/ipython_1762789338226/work
ipython_pygments_lexers @ file:///croot/ipython_pygments_lexers_1744753235686/work
jedi @ file:///croot/jedi_1733987392413/work
Jinja2==3.1.6
joblib==1.5.3
jupyter_client @ file:///home/task_176839840248915/croot/jupyter_client_1768399002805/work
jupyter_core @ file:///home/task_176458820445830/conda-bld/jupyter_core_1764588284850/work
MarkupSafe==3.0.3
matplotlib-inline @ file:///home/task_176277916602646/conda-bld/matplotlib-inline_1762779180318/work
moviepy==2.2.1
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.1
multiprocess==0.70.16
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openpyxl==3.1.5
packaging @ file:///home/task_176104874243446/conda-bld/packaging_1761049080023/work
pandas==3.0.1
parso @ file:///home/task_176278181429383/conda-bld/parso_1762781859742/work
peft==0.18.1
pexpect @ file:///home/task_176253513024890/conda-bld/pexpect_1762535937809/work
pillow==11.3.0
platformdirs @ file:///home/task_176235638281825/conda-bld/platformdirs_1762356487196/work
proglog==0.1.12
prompt_toolkit @ file:///home/task_176174483920045/conda-bld/prompt-toolkit_1761744853185/work
propcache==0.4.1
protobuf==7.34.0
psutil @ file:///home/task_176189642211680/conda-bld/psutil_1761896463664/work
ptyprocess @ file:///opt/miniconda3/conda-bld/ptyprocess_1762424170819/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=3470be7f810474c8a2ecfcd6e02acc6aea8483ab595417fa4e336362a349933e
pure_eval @ file:///home/task_175706684826776/conda-bld/pure_eval_1757067049051/work
pyarrow==23.0.1
Pygments @ file:///home/task_176243120367123/conda-bld/pygments_1762431407417/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq @ file:///home/task_176237559504765/conda-bld/pyzmq_1762375608075/work
qwen-vl-utils==0.0.14
regex==2026.2.28
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
setuptools==80.10.2
six @ file:///croot/six_1744271502820/work
stack_data @ file:///home/task_175706679004827/conda-bld/stack_data_1757067036633/work
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torchao==0.16.0
torchvision==0.25.0
tornado @ file:///home/task_176599247062576/croot/tornado_1765992645217/work
tqdm==4.67.3
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.57.6
triton==3.6.0
trl==0.24.0
typeguard==4.5.1
typing_extensions==4.15.0
tyro==1.0.8
unsloth==2026.2.1
unsloth_zoo==2026.2.1
urllib3==2.6.3
wcwidth @ file:///home/task_176777912523564/croot/wcwidth_1767779146734/work
wheel==0.46.3
xformers==0.0.35
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
